+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.4.9
+
+### Patch Changes
+
+-   Updated dependencies [d549173]
+    -   @sap-ux/project-access@1.25.7
+
 ## 0.4.8
 
 ### Patch Changes